+++ CHANGELOG.md
@@ -2,10 +2,11 @@ Changelog
 =========
 
 NOTE: isort follows the [semver](https://semver.org/) versioning standard.
-### 5.0.5 July 6, 2020
+### 5.0.5 July 7, 2020
   - Fixed #1285: packaging issue with bundling tests via poetry.
   - Fixed #1284: Regression when sorting `.pyi` files from CLI using black profile.
   - Fixed #1275 & #1283: Blank line after docstring removed.
+  - Fixed #1298: CLI Help out of date with isort 5.
 
 ### 5.0.4 July 6, 2020
   - Fixed #1264: a regression with comment handling and `force_sort_within_sections` config option